CircuitSimulator Refactor (#12)
Refactor the CircuitSimulator to make extensibility easier for new simulation types
