+++ .gitignore
@@ -83,5 +83,8 @@ docs/sphinx/_doxygen
 docs/sphinx/_mdgen
 **/_build/*
 
+# third party integrations
+simulators/
+
 # macOS
 .DS_Store
