+++ runtime/nvqir/NVQIR.cpp
@@ -341,6 +341,7 @@ Result *__quantum__qis__measure__body(Array *pauli_arr, Array *qubits) {
   // Some backends may better handle the observe task.
   // Let's give them that opportunity.
   if (currentContext->canHandleObserve) {
+    circuitSimulator->flushGateQueue();
     auto result = circuitSimulator->observe(*currentContext->spin.value());
     currentContext->expectationValue = result.expectationValue;
     currentContext->result = cudaq::sample_result(result);
@@ -377,6 +378,7 @@ Result *__quantum__qis__measure__body(Array *pauli_arr, Array *qubits) {
     }
   }
 
+  circuitSimulator->flushGateQueue();
   int shots = 0;
   if (currentContext->shots > 0) {
     shots = currentContext->shots;
@@ -399,6 +401,7 @@ Result *__quantum__qis__measure__body(Array *pauli_arr, Array *qubits) {
         circuitSimulator->rx(angle, it->second);
       }
     }
+    circuitSimulator->flushGateQueue();
   }
 
   return ResultZero;
